refactor: make presentation tier does not depend on infra tier
