+++ src/app.module.ts
@@ -1,17 +1,10 @@
 import { Module } from '@nestjs/common';
 import { UserController } from './user.controller';
-import { UserRepository } from './domain/user.repository';
-import { OnMemoryUserRepository } from './infra/on-memory/user.repository';
-import { UserAppService } from './application/user.service';
-
-export const userRepositoryProvider = {
-  provide: UserRepository,
-  useClass: OnMemoryUserRepository,
-};
+import { UserModule } from './application/user.module';
 
 @Module({
-  imports: [],
+  imports: [UserModule],
   controllers: [UserController],
-  providers: [UserAppService, userRepositoryProvider],
+  providers: [],
 })
 export class AppModule {}
