+++ src/application/user.module.ts
@@ -1,0 +1,17 @@
+import { Module } from '@nestjs/common';
+import { UserRepository } from '../domain/user.repository';
+import { OnMemoryUserRepository } from '../infra/on-memory/user.repository';
+import { UserAppService } from './user.service';
+
+export const userRepositoryProvider = {
+  provide: UserRepository,
+  useClass: OnMemoryUserRepository,
+};
+
+@Module({
+  imports: [],
+  controllers: [],
+  providers: [UserAppService, userRepositoryProvider],
+  exports: [UserAppService],
+})
+export class UserModule {}
